scheduler/requirements.txt
describe-it==2.1.0
ipdb==0.8.1
ipython==3.1.0
nose==1.3.7
nose-terse-output==1.1.0
PyHamcrest==1.8.3
pyreadline==2.0
six==1.9.0
